+++ NEWS.md
@@ -8,6 +8,8 @@ that attempt to access a Databricks cluster are successful (#166)
 
 ### Improvements
 
+* Adds support for Databricks Viewer OAuth credentials. 
+
 * Adds support for Snowflake's [Snowpark Connect](https://docs.snowflake.com/en/developer-guide/snowpark-connect/snowpark-connect-overview).
 New method name is `snowpark_connect`.